♻️ refactor: add check functions 입력한 값이 유효한 값인지에 대해서 확인하는 함수를 분리
